+++ src/common/Utils.ts
@@ -8,7 +8,8 @@ import * as vscode from "vscode";
 import { EXTENSION_ID, EXTENSION_NAME, SETTINGS_EXPERIMENTAL_STORE_NAME } from "../client/constants";
 import { CUSTOM_TELEMETRY_FOR_POWER_PAGES_SETTING_NAME } from "./OneDSLoggerTelemetry/telemetryConstants";
 import { PacWrapper } from "../client/pac/PacWrapper";
-import { AUTH_CREATE_FAILED, AUTH_CREATE_MESSAGE, PAC_SUCCESS } from "./copilot/constants";
+import { AUTH_CREATE_FAILED, AUTH_CREATE_MESSAGE, DataverseEntityNameMap, EntityFieldMap, FieldTypeMap, PAC_SUCCESS } from "./copilot/constants";
+import { IActiveFileData, IActiveFileParams } from "./copilot/model";
 
 export function getSelectedCode(editor: vscode.TextEditor): string {
     if (!editor) {
@@ -151,4 +152,30 @@ export async function createAuthProfileExp(pacWrapper: PacWrapper | undefined) {
         vscode.window.showErrorMessage(AUTH_CREATE_FAILED);
         return;
     }
+}
+
+export function getActiveEditorContent(): IActiveFileData {
+    const activeEditor = vscode.window.activeTextEditor;
+    const activeFileData: IActiveFileData = {
+        activeFileContent: '',
+        activeFileParams: {
+            dataverseEntity: '',
+            entityField: '',
+            fieldType: ''
+        } as IActiveFileParams
+    };
+    if (activeEditor) {
+        const document = activeEditor.document;
+        const fileName = document.fileName;
+        const relativeFileName = vscode.workspace.asRelativePath(fileName);
+
+        const activeFileParams: string[] = getLastThreePartsOfFileName(relativeFileName);
+
+        activeFileData.activeFileContent = document.getText();
+        activeFileData.activeFileParams.dataverseEntity = DataverseEntityNameMap.get(activeFileParams[0]) || "";
+        activeFileData.activeFileParams.entityField = EntityFieldMap.get(activeFileParams[1]) || "";
+        activeFileData.activeFileParams.fieldType = FieldTypeMap.get(activeFileParams[2]) || "";
+    }
+
+    return activeFileData;
 }
